data_scripts/pdf_to_csv/requirements.txt
tabula-py==1.0.0
